feat: 게임 리소스 권한 체크 로직 적용(#13)

+++ core/core-api/src/main/java/org/ject/recreation/core/domain/game/GameService.java
@@ -14,6 +14,7 @@ import org.ject.recreation.core.domain.game.question.Question;
 import org.ject.recreation.core.domain.game.question.QuestionReader;
 import org.ject.recreation.core.domain.game.question.QuestionResult;
 import org.ject.recreation.core.support.error.CoreException;
+import org.ject.recreation.core.support.error.ErrorData;
 import org.ject.recreation.core.support.error.ErrorType;
 import org.ject.recreation.storage.db.core.*;
 
@@ -24,6 +25,8 @@ import java.util.List;
 import java.util.UUID;
 import java.util.stream.IntStream;
 import java.util.stream.Collectors;
+
+import static org.ject.recreation.core.support.error.ErrorType.GAME_FORBIDDEN;
 
 @Service
 @RequiredArgsConstructor
@@ -62,7 +65,7 @@ public class GameService {
 
         return new GameDetailResult(
                 game.gameTitle(),
-                game.nickname(),
+                game.creatorNickname(),
                 game.questionCount(),
                 game.version(),
                 questions.stream()
@@ -81,9 +84,12 @@ public class GameService {
         return generatePresignedUrls(UUID.randomUUID(), presignedUrlQuery);
     }
 
-    public PresignedUrlListResult getPresignedUrls(UUID gameId, PresignedUrlQuery presignedUrlQuery) {
+    public PresignedUrlListResult getPresignedUrls(String curUserEmail, UUID gameId, PresignedUrlQuery presignedUrlQuery) {
         Game game = gameReader.getGameByGameId(gameId);
-        // TODO: 게임 권한 소지 여부 확인 로직 추가
+
+        if (!game.creatorEmail().equals(curUserEmail)) {
+            throw new CoreException(GAME_FORBIDDEN, ErrorData.of("gameId", gameId));
+        }
 
         return generatePresignedUrls(gameId, presignedUrlQuery);
     }
@@ -140,6 +146,10 @@ public class GameService {
         GameEntity existingGame = gameRepository.findById(gameId)
                 .orElseThrow(() -> new CoreException(ErrorType.GAME_NOT_FOUND));
 
+        if (!existingGame.getGameCreator().getEmail().equals(userInfo.getEmail())) {
+            throw new CoreException(GAME_FORBIDDEN, ErrorData.of("gameId", gameId));
+        }
+
         // 기존 질문들 삭제
         questionRepository.deleteByGame(existingGame);
 
@@ -181,25 +191,34 @@ public class GameService {
     }
 
     @Transactional
-    public void deleteGame(UUID gameId) {
+    public void deleteGame(String curUserEmail, UUID gameId) {
         Game game = gameReader.getGameByGameId(gameId);
-        // TODO: 게임 권한 소지 여부 확인 로직 추가
+
+        if (!game.creatorEmail().equals(curUserEmail)) {
+            throw new CoreException(GAME_FORBIDDEN, ErrorData.of("gameId", gameId));
+        }
 
         gameWriter.deleteGame(game);
     }
 
     @Transactional
-    public void shareGame(UUID gameId) {
+    public void shareGame(String curUserEmail, UUID gameId) {
         Game game = gameReader.getGameByGameId(gameId);
-        // TODO: 게임 권한 소지 여부 확인 로직 추가
+
+        if (!game.creatorEmail().equals(curUserEmail)) {
+            throw new CoreException(GAME_FORBIDDEN, ErrorData.of("gameId", gameId));
+        }
 
         gameWriter.shareGame(game);
     }
 
     @Transactional
-    public void unShareGame(UUID gameId) {
+    public void unShareGame(String curUserEmail, UUID gameId) {
         Game game = gameReader.getGameByGameId(gameId);
-        // TODO: 게임 권한 소지 여부 확인 로직 추가
+
+        if (!game.creatorEmail().equals(curUserEmail)) {
+            throw new CoreException(GAME_FORBIDDEN, ErrorData.of("gameId", gameId));
+        }
 
         gameWriter.unShareGame(game);
     }
